requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-mail==1.4.2
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyodbc==5.2.0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlalchemy-filterset==2.3.0
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
websocket==0.2.1
websockets==14.1
zope.event==5.0
zope.interface==7.2
